+++ avilib/avilib.c
@@ -466,7 +466,8 @@ static int avi_add_odml_index_entry(avi_t *AVI, const unsigned char *tag,
 	    fprintf (stderr, "Internal error in avilib - redefine NR_IXNN_CHUNKS\n");
 	    fprintf (stderr, "[avilib dump] cur_std_idx=%d NR_IXNN_CHUNKS=%d"
 		    "POS=%lld towrite=%lld\n",
-		    cur_std_idx,NR_IXNN_CHUNKS, AVI->pos, towrite);
+		    cur_std_idx,NR_IXNN_CHUNKS,
+		    (unsigned long long)AVI->pos, (unsigned long long)towrite);
 	    return -1;
 	}
 
@@ -3330,7 +3331,8 @@ long AVI_read_audio(avi_t *AVI, char *audbuf, long bytes)
       xio_lseek(AVI->fdes, pos, SEEK_SET);
       if ( (ret = avi_read(AVI->fdes,audbuf+nr,todo)) != todo)
       {
-	 fprintf(stderr, "XXX pos = %lld, ret = %lld, todo = %ld\n", pos, ret, todo);
+	 fprintf(stderr, "XXX pos = %lld, ret = %lld, todo = %ld\n",
+                     (unsigned long long)pos, (unsigned long long)ret, todo);
          AVI_errno = AVI_ERR_READ;
          return -1;
       }
